% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinp.R
\name{combinp}
\alias{combinp}
\title{Generate corrected p-value based on p-value and logFC in the expression matrix}
\usage{
combinp(node_attr = NULL)
}
\arguments{
\item{node_attr}{logFC and p value in the two columns}
}
\value{
the matrix of two columns: gene, weight(corrected p-value)
}
\description{
based on the formula: corrected p-value = 2*( 1-pnorm( (-log10( p-value )) * abs( log2FC )) ), corrected p-value was generated
}
\examples{
data("node_attr_Example")
result <- combinp(node_attr_Example)

}
\references{
Hongbo Shi, Jiayao Li, Qiong Song et al. (2019) Systematic identification and analysis of dysregulated miRNA and transcription factor feed-forward loops in hypertrophic cardiomyopathy
}
