% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largestComp.R
\name{largestComp}
\alias{largestComp}
\title{Extract largest component of network}
\usage{
largestComp(network)
}
\arguments{
\item{network}{A graph in igraph format}
}
\value{
A new graph object that represents the largest component of the given network.
}
\description{
The function extracts the largest component of a network.
}
\examples{
data("network_Example")
largestComp(network_Example)

}
