% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncRNAcircRNA-package.R
\docType{package}
\name{ncRNAcircRNA-package}
\alias{ncRNAcircRNA-package}
\alias{ncRNAcircRNA}
\title{ncRNAcircRNA: Search for modules in a node-weighted ncRNA network}
\description{
ncRNAcircRNA constructs a node-weighted ncRNA network, performs module searching, generates simulation data from random networks,
normalizes module scores using simulation data, removes un-qualified modules, and orders resultant modules according to their significance.
}
\details{
This package takes three types of data as input: a list of genes with
association p-values and logFC, a human ncRNA network. generate_graph constructs a
node-weighted ncRNA network. runmodule performs  module
search upon the node-weighted ncRNA network.
}
\references{


Hongbo Shi, Jiayao Li, Qiong Song et al. (2019) Systematic identification and analysis of dysregulated miRNA and transcription factor feed-forward loops in hypertrophic cardiomyopathy

Peilin Jia, Siyuan Zheng, Jirong Rong, Wei Zheng, Zhongming Zhao. (2011) Bioformatics. dmGWAS: dense module searching for genome-wide association studies in protein-protein interaction networks.
}
