% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_Example.R
\docType{data}
\name{network_Example}
\alias{network_Example}
\title{Example network for aggregation statistics}
\usage{
data("network_Example")
}
\description{
Data example consisting of a matrix of network.
}
\examples{
data("network_Example")
network_Example[1:100,]
}
