% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_attr_Example.R
\docType{data}
\name{node_attr_Example}
\alias{node_attr_Example}
\title{Example node_attr for aggregation statistics}
\usage{
data("node_attr_Example")
}
\description{
Data example consisting of a matrix of logFC, p-values. Each gene has corresponding logFC, p-values.
}
\examples{
data("node_attr_Example")
node_attr_Example[1:100,]
}
