% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runmodule.R
\name{runmodule}
\alias{runmodule}
\title{Run module search function}
\usage{
runmodule(network, gene2weight, d = 2, r = 0.1, method = c("Heinz", "GS"))
}
\arguments{
\item{network}{A data frame containing a symbolic edge list of the ncRNA
network}

\item{gene2weight}{weight A data frame containing two columns: the first is unique
gene identifier (should be coordinate with the node symbol used in ncRNA network);
the second is gene-based corrected p-value derived from differentially gene analysis}

\item{d}{An integer used to define the order of neighbour genes to be
searched if the method is "GS". This parameter is always set up as 2}

\item{r}{A float indicating the cut-off for increment during module expanding
process if the method is "GS". Greater r will generate smaller module. Default is 0.1.}

\item{method}{a character string indicating which the search method is to be computed
. One of "Heinz" (default), "GS": can be abbreviated}
}
\value{
\code{runmodule} returns a list containing relevant data and results,
  including:

 \tabular{ll}{
   \code{GWPI} \tab the edge-weighted network used for searching \cr
   \code{graph.g.weight} \tab the edge-weighted of each gene in the edge-weighted network \cr
   \code{module} \tab list of genes comprising each module, named for the seed gene \cr
   \code{module.score.matrix} \tab contains Zm, Zn and Zcount \cr
   \code{Zi} \tab contains Zn \cr
 }
}
\description{
runmodule constructs a node-weighted ncRNA network, performs module searching, generates simulation data from random networks,

normalizes module scores using simulation data, removes un-qualified modules, and orders resultant modules according to their significance.
}
\examples{
\dontrun{
 data("node_attr_Example")
 data("network_Example")
 geneweight <- combinp(node_attr_Example)
 gene2weight = gene2weight[1:100,]
 res.list <- runmodule(network = network_Example, gene2weight, method = "Heinz")
}

}
\references{
Hongbo Shi, Jiayao Li, Qiong Song et al. (2019) Systematic identification and analysis of dysregulated miRNA and transcription factor feed-forward loops in hypertrophic cardiomyopathy

Peilin Jia, Siyuan Zheng, Jirong Rong, Wei Zheng, Zhongming Zhao. (2011) Bioformatics. dmGWAS: dense module searching for genome-wide association studies in protein-protein interaction networks.
}
