% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subNetwork_only.R
\name{subNetwork_only}
\alias{subNetwork_only}
\title{Create a subGraph}
\usage{
subNetwork_only(nodeList, network)
}
\arguments{
\item{nodeList}{Character vector of nodes, contained in the subgraph.}

\item{network}{Graph that is used for subgraph extraction}
}
\value{
A graph object.
}
\description{
The function creates a subgraph with the nodes given in the nodeList
}
\examples{
library(igraph)
edgel <- cbind(c("1", "2", "3", "4", "5", "6", "7"), c("b", "c", "d", "e", "f", "a", "b"))
g <- graph.edgelist(edgel, directed=TRUE)
node.list <- c("a", "b", "c", "1")
graph <- subNetwork_only(nodeList=node.list, network=g)


}
